requirements.txt
librosa
pydub
fuzzysearch
python-Levenshtein
numba==0.48
google-cloud-speech
numpy
